Support sending ICP to principals + highlight when destination account invalid
